fixup! Core: Don't reset snapshotLog in TableMetadata.removeRef
